made everything double instead of float with precision

+++ app/models.py
@@ -42,11 +42,11 @@ class UserMapStats(db.Model):
     map_id = db.Column(db.Integer, db.ForeignKey("maps.id"), nullable=False)
     total_time = db.Column(db.Integer, nullable=False, default=0)
     total_score = db.Column(db.Integer, nullable=False, default=0)
-    total_distance = db.Column(db.Float(precision=53), nullable=False, default=0)
+    total_distance = db.Column(db.Double, nullable=False, default=0)
     total_guesses = db.Column(db.Integer, nullable=False, default=0)
 
-    high_average_score = db.Column(db.Integer, nullable=False, default=0)
-    high_average_distance = db.Column(db.Float(precision=53), nullable=False, default=0)
+    high_average_score = db.Column(db.Float, nullable=False, default=0)
+    high_average_distance = db.Column(db.Double, nullable=False, default=0)
     high_average_time = db.Column(db.Integer, nullable=False, default=0)
     high_round_number = db.Column(db.Integer, nullable=False, default=0)
 
@@ -57,10 +57,10 @@ class Guess(db.Model):
     user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
     round_id = db.Column(db.Integer, db.ForeignKey("rounds.id"), nullable=False)
 
-    latitude = db.Column(db.Float(precision=53), nullable=False)
-    longitude = db.Column(db.Float(precision=53), nullable=False)
+    latitude = db.Column(db.Double, nullable=False)
+    longitude = db.Column(db.Double, nullable=False)
 
-    distance = db.Column(db.Float(precision=53),nullable=False)
+    distance = db.Column(db.Double,nullable=False)
     score = db.Column(db.Integer, nullable=False)
     time = db.Column(db.Integer, nullable=False, default=0)
 
@@ -87,8 +87,8 @@ class SVLocation(db.Model):
     __tablename__ = "svlocations"
 
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    latitude = db.Column(db.Float(precision=53), nullable=False)
-    longitude = db.Column(db.Float(precision=53), nullable=False)
+    latitude = db.Column(db.Double, nullable=False)
+    longitude = db.Column(db.Double, nullable=False)
 
     rounds = db.relationship("Round", backref="location", cascade="all,delete")
 
@@ -155,7 +155,7 @@ class RoundStats(db.Model):
     round = db.Column(db.Integer, nullable=False)
     total_time = db.Column(db.Integer, nullable=False, default=0)
     total_score = db.Column(db.Integer, nullable=False, default=0)
-    total_distance = db.Column(db.Float(precision=53), nullable=False, default=0)
+    total_distance = db.Column(db.Double, nullable=False, default=0)
 
 ############################################################################################
 #   MAP                                                                                    #
@@ -168,13 +168,13 @@ class GameMap(db.Model):
     uuid = db.Column(db.String(36), default=lambda: str(uuid.uuid4()), unique=True)
     creator_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
     
-    start_latitude = db.Column(db.Float(precision=53), nullable=False,default=-1)
-    start_longitude = db.Column(db.Float(precision=53), nullable=False,default=-1)
-    end_latitude = db.Column(db.Float(precision=53), nullable=False,default=-1)
-    end_longitude = db.Column(db.Float(precision=53), nullable=False,default=-1)
+    start_latitude = db.Column(db.Double, nullable=False,default=-1)
+    start_longitude = db.Column(db.Double, nullable=False,default=-1)
+    end_latitude = db.Column(db.Double, nullable=False,default=-1)
+    end_longitude = db.Column(db.Double, nullable=False,default=-1)
     
     total_weight = db.Column(db.Integer, nullable=False, default=0)
-    max_distance = db.Column(db.Float(precision=53), nullable=False,default=-1)
+    max_distance = db.Column(db.Double, nullable=False,default=-1)
     
     map_bounds = db.relationship("MapBound", backref="map", cascade="all,delete")
     sessions = db.relationship("Session", backref="map", cascade="all,delete")
@@ -192,7 +192,7 @@ class MapStats(db.Model):
     map_id = db.Column(db.Integer, db.ForeignKey("maps.id"), nullable=False)
     total_time = db.Column(db.Integer, nullable=False, default=0)
     total_score = db.Column(db.Integer, nullable=False, default=0)
-    total_distance = db.Column(db.Float(precision=53), nullable=False, default=0)
+    total_distance = db.Column(db.Double, nullable=False, default=0)
     total_guesses = db.Column(db.Integer, nullable=False, default=0)
     
     total_generation_time = db.Column(db.Integer, nullable=False, default=0)
@@ -203,10 +203,10 @@ class Bound(db.Model):
     
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     
-    start_latitude = db.Column(db.Float(precision=53), nullable=False)
-    start_longitude = db.Column(db.Float(precision=53), nullable=False)
-    end_latitude = db.Column(db.Float(precision=53), nullable=False)
-    end_longitude = db.Column(db.Float(precision=53), nullable=False)
+    start_latitude = db.Column(db.Double, nullable=False)
+    start_longitude = db.Column(db.Double, nullable=False)
+    end_latitude = db.Column(db.Double, nullable=False)
+    end_longitude = db.Column(db.Double, nullable=False)
     
     map_bounds = db.relationship("MapBound", backref="bound", cascade="all,delete")
